refactor: component folder pattern
