+++ src/components/App/App.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
-import logo from './logo.svg';
-import { Counter } from './features/counter/Counter';
-import './App.css';
+import React from "react";
+import logo from "./logo.svg";
+import { Counter } from "../../features/counter/Counter";
+import "./App.css";
 
-function App() {
+export function App() {
   return (
     <div className="App">
       <header className="App-header">
@@ -54,5 +54,3 @@ function App() {
     </div>
   );
 }
-
-export default App;
